+++ src/components/organisms/ArtistList.tsx
@@ -3,7 +3,7 @@ import Artist from "@/service/artist/Artist";
 import { ArtistData, SearchProps } from "@/types/_type";
 import React, { useEffect, useState } from "react";
 import useSWR, { mutate } from "swr";
-import ArtistWrap from "../ArtistWrap";
+import ArtistWrap from "../templetes/ArtistWrap";
 import Pagination from "../atoms/Pagination";
 import Loading from "../common/Loading";
